refactor new commission page into StudioLive
